Fibonacci

Description

Documentation

fibb :: Int -> Integer

Binet Fibonacci.

Uses the Binet Formula to generate the nth Fibonacci number.

fibi :: Int -> Integer

Fibonacci Index.

This retrieves the nth Fibonacci value by index from a sequence produced by fibs.

fibp :: Int -> Integer

Parallel Fibonacci.

Calculate Fibonacci in parallel.

From: "Practical Concurrent Haskell: With Big Data Applications", Chapter 3, Parallelism and Concurrency with Haskell

fibr :: Int -> Integer

Fibonacci using code from Rebecca Skinner.

From: "Effective Haskell" by Rebecca Skinner Chapter 2, Hands-On With Infinite Fibonacci Numbers

fibs :: [Integer]

Fibonacci Sequence.

The fibs function is a definition of an infinite list of Fibonacci numbers. It is defined recursively, by starting with the two numbers 0 and 1, and then generating the rest of the sequence by adding adjacent pairs of numbers in the list.

The zipWith function is used to create a list where each element is the sum of the previous two elements in the list. Specifically, zipWith (+) fibs (tail fibs) takes two lists as input: fibs, which is the original sequence starting with 0 and 1, and tail fibs, which is the same sequence but with the first element removed. It applies the + function to each corresponding pair of elements in the two lists, resulting in a new list where each element is the sum of the previous two elements in fibs.

The resulting list is then prepended with 0 and 1 to create the full sequence, which is an infinite list of Fibonacci numbers [0, 1, 1, 2, 3, 5, 8, 13, 21, ...].

This is very similar to the fibr function, but it uses zipWith.

fibt :: Int -> Integer

Traditional Fibonacci.

The traditional recursive Fibonacci.